% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities.R
\name{bin_expression}
\alias{bin_expression}
\title{Bin Expression to 256 Bins}
\usage{
bin_expression(featuresdf)
}
\arguments{
\item{featuresdf}{a data frame containing expression information}
}
\value{
a character vector containing hexcoded color with a length the same
as the row number of the input data frame
}
\description{
\code{bin_expression()} takes a data frame and bin every column to 256 bins,
and generate Hex-coded color code according to the composite of the columns.
It takes three or less columns.
}
