% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities.R
\name{make_legend}
\alias{make_legend}
\title{Extract Figure Legend from Individual Plots for Later Integration to the
Main Plot}
\usage{
make_legend(featurekey, plotdf)
}
\arguments{
\item{featurekey}{a character vector containing the features prefixed by
keys (if using Seuratv3) to be plotted}

\item{plotdf}{a data frame containing cell embedings, expression level, and
color codes genearted by \code{\link{bin_expression}()}}
}
\value{
a ggplot2 object
}
\description{
Since \code{ggplot2} does not support superimposing multiple
\code{geom_point()} with different color codes,
\code{\link{MultiFeaturePlot}()} bypasses the limitation by generating one
custom color scale integrating multiple expression data. To get legends for
that, individual \code{geom_point()} ggplots are generated here, and
extracted for later integration with the main plot.
}
