% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_main.R
\name{MultiFeaturePlot}
\alias{MultiFeaturePlot}
\title{Plotting Multiple Features on a Dimensionality Reduced Space}
\usage{
MultiFeaturePlot(object, features, assay = "RNA", pt.size = 1,
  reduction = "tsne", cells.use = NULL, title = NULL)
}
\arguments{
\item{object}{a Seurat object (v2 or v3)}

\item{features}{a character vector of features to plot}

\item{assay}{a character string indicating the assay to plot (SeuratV3 only)}

\item{pt.size}{a number indicating the size of points on the plot}

\item{reduction}{a character string indicating the reduction to embed the
cells on the plot (e.g., "tsne" or "pca")}

\item{cells.use}{a character vector of cell names indicating the cells to
include in the plot (try \code{\link[Seurat]{WhichCells}()})}

\item{title}{a character string for customized plot title. If not set, the
default is the feature names}
}
\value{
a ggplot object
}
\description{
\code{MultiFeaturePlot()} plots cells in a Seurat object into a t-SNE space
or other dimensionality reduced space for visualization. Each point will be
colored according to log-normalized expression level (from \code{@data} slot
of a Seurat object). The first gene will be colored in red, the second gene
will be colored in green, and the third gene will be colored in blue.
}
\details{
Specifically, the gene expression is translated into color by first taking
the full range of expression from zero to highest, binning them into 256
bins, and use the binned expression level as 8-bit color code.
}
